Cache clear optimization

+++ GSPlayer/Classes/Cache/VideoCacheManager.swift
@@ -40,16 +40,30 @@ public enum VideoCacheManager {
         }
     }
     
-    public static func cleanAllCache() throws {
+    public static func cleanAllCache(excepts urls: [URL]? = nil) throws {
         let fileManager = FileManager.default
         let fileContents = try fileManager.contentsOfDirectory(atPath: directory)
+        var excepts = urls?.map{cachedFilePath(for: $0)} ?? []
+        for filePath in excepts {
+            excepts.append(VideoCacheConfiguration.configurationFilePath(for: filePath))
+        }
         
         for fileContent in fileContents {
             let filePath = directory.appendingPathComponent(fileContent)
-            try fileManager.removeItem(atPath: filePath)
+            if excepts.isEmpty || !excepts.contains(filePath) {
+                try fileManager.removeItem(atPath: filePath)
+            }
         }
     }
     
+    public static func deleteFile(for url: URL) throws {
+        let fileManager = FileManager.default
+        let filePath = cachedFilePath(for: url)
+        if fileManager.fileExists(atPath: filePath) {
+            try fileManager.removeItem(atPath: filePath)
+            try fileManager.removeItem(atPath: VideoCacheConfiguration.configurationFilePath(for: filePath))
+        }
+    }
 }
 
 @objc public class VideoCacheManagerUtils : NSObject {
@@ -61,7 +75,11 @@ public enum VideoCacheManager {
         VideoCacheManager.calculateCachedSize();
     }
     
-    @objc public static func cleanAllCache() throws {
-        try VideoCacheManager.cleanAllCache();
+    @objc public static func cleanAllCache(excepts urls: [URL]? = nil) throws {
+        try VideoCacheManager.cleanAllCache(excepts: urls);
+    }
+    
+    @objc public static func deleteFile(for url: URL) throws {
+        try VideoCacheManager.deleteFile(for: url);
     }
 }
